+++ examples/interception.rs
@@ -2,8 +2,8 @@ use std::sync::Arc;
 
 use base64::prelude::BASE64_STANDARD;
 use base64::Engine;
-use chaser_oxide::browser::{Browser, BrowserConfig};
-use chaser_oxide::cdp::browser_protocol::fetch::{
+use ignition::browser::{Browser, BrowserConfig};
+use ignition::cdp::browser_protocol::fetch::{
     ContinueRequestParams, EventRequestPaused, FulfillRequestParams,
 };
 use futures::StreamExt;