+++ examples/storage-cookie.rs
@@ -1,6 +1,6 @@
-use chaser_oxide::browser::Browser;
-use chaser_oxide::browser::BrowserConfig;
-use chaser_oxide::cdp::browser_protocol::network::CookieParam;
+use ignition::browser::Browser;
+use ignition::browser::BrowserConfig;
+use ignition::cdp::browser_protocol::network::CookieParam;
 use futures::StreamExt;
 
 #[tokio::main]